+++ forui/lib/src/widgets/accordion/accordion_item.dart
@@ -3,7 +3,6 @@ import 'package:flutter/widgets.dart';
 
 import 'package:forui/forui.dart';
 import 'package:forui/src/widgets/accordion/accordion.dart';
-import 'package:forui/src/widgets/accordion/accordion_control.dart';
 import 'package:forui/src/widgets/accordion/accordion_controller.dart';
 
 /// A marker interface which denotes that mixed-in widgets can be used in a [FAccordion].