+++ forui/lib/widgets/accordion.dart
@@ -7,5 +7,5 @@ library forui.widgets.accordion;
 
 export '../src/widgets/accordion/accordion.dart' hide InheritedAccordionData;
 export '../src/widgets/accordion/accordion_control.dart' show FAccordionControl;
-export '../src/widgets/accordion/accordion_controller.dart' hide InternalAccordionController;
+export '../src/widgets/accordion/accordion_controller.dart' hide InternalAccordionController, LiftedController;
 export '../src/widgets/accordion/accordion_item.dart';